+++ package.json
@@ -54,10 +54,10 @@
     "@prisma/adapter-pg": "^7.10.0",
     "@prisma/client": "^7.10.0",
     "@svgr/webpack": "^8.1.0",
-    "@tanstack/charts": "^0.15.0",
-    "@tanstack/charts-scales": "^0.15.0",
-    "@tanstack/react-charts": "^0.15.0",
-    "@tanstack/react-query": "^5.102.6",
+    "@tanstack/charts": "^0.16.0",
+    "@tanstack/charts-scales": "^0.16.0",
+    "@tanstack/react-charts": "^0.16.0",
+    "@tanstack/react-query": "^5.102.8",
     "@tanstack/react-table": "^9.2.3",
     "classnames": "^2.5.1",
     "d3-shape": "^3.2.0",
@@ -90,7 +90,7 @@
     "@types/d3-shape": "^3.2.0",
     "@types/jest": "^30.0.0",
     "@types/js-cookie": "^3.0.6",
-    "@types/node": "^26.3.0",
+    "@types/node": "^26.4.0",
     "@types/pg": "catalog:",
     "@types/react": "^19.2.18",
     "@types/react-dom": "^19.2.5",
