+++ src/lib/private-route-response.ts
@@ -1,20 +1,15 @@
 import { NextResponse } from "next/server";
+import { applyPrivateRouteCacheHeaders } from "src/lib/private-route-cache";
 
-/** Prevent CDN/proxy caching of cookie-authenticated API responses. */
-export const PRIVATE_ROUTE_CACHE_HEADERS = {
-  "Cache-Control":
-    "private, no-store, no-cache, must-revalidate, proxy-revalidate",
-  Pragma: "no-cache",
-  Vary: "Cookie",
-} as const;
+export {
+  isPrivateSessionApiRoute,
+  PRIVATE_ROUTE_CACHE_HEADERS,
+} from "src/lib/private-route-cache";
 
 export const withPrivateRouteHeaders = (
   response: NextResponse,
 ): NextResponse => {
-  for (const [key, value] of Object.entries(PRIVATE_ROUTE_CACHE_HEADERS)) {
-    response.headers.set(key, value);
-  }
-
+  applyPrivateRouteCacheHeaders(response.headers);
   return response;
 };
 
@@ -30,22 +25,7 @@ export const privateRouteJson = <T>(
   init?: ResponseInit,
 ): NextResponse => {
   const headers = new Headers(init?.headers);
-
-  for (const [key, value] of Object.entries(PRIVATE_ROUTE_CACHE_HEADERS)) {
-    headers.set(key, value);
-  }
+  applyPrivateRouteCacheHeaders(headers);
 
   return NextResponse.json(body, { ...init, headers });
-};
-
-export const isPrivateSessionApiRoute = (pathname: string): boolean => {
-  if (pathname.startsWith("/api/auth")) {
-    return true;
-  }
-
-  if (!pathname.startsWith("/api/crates")) {
-    return false;
-  }
-
-  return !pathname.startsWith("/api/crates/public");
 };
